[step4] content length int타입으로 읽도록 변경, 헤더, 바디 읽는 메서드 분리

+++ src/main/java/webserver/common/HttpHeader.java
@@ -25,12 +25,12 @@ public class HttpHeader{
         this.headers = headers;
     }
 
-    public long getContentLength() {
+    public int getContentLength() {
         String contentLength = getFirst(CONTENT_LENGTH);
-        return contentLength != null ? Long.parseLong(contentLength) : -1L;
+        return contentLength != null ? Integer.parseInt(contentLength) : -1;
     }
 
-    public void setContentLength(long contentLength) {
+    public void setContentLength(int contentLength) {
         set(CONTENT_LENGTH, String.valueOf(contentLength));
     }
